Crate azalea_registry

Azalea Registry

Minecraft has a concept called “registries”, which are primarily used to determine ids in the protocol. The contents of this crate are automatically generated using Minecraft’s built-in data generator.

Modules

tags

Structs

OptionalRegistry

A registry that might not be present. This is transmitted as a single varint in the protocol.

Enums

Activity

The AI code that’s currently being executed for the entity.

ArmorMaterial

Attribute

BannerPattern

Block

An enum of every type of block in the game. To represent a block state, use azalea_block::BlockState or the azalea_block::Block trait.

BlockEntityKind

An enum that contains every type of block entity. A block entity is a block that contains data that can’t be represented as just a block state, like how chests store items.

BlockKind

BlockPredicateKind

CatVariant

ChunkStatus

CommandArgumentKind

ConsumeEffectKind

CreativeModeTab

CustomRegistry

A registry that will either take an ID or a resource location.

CustomStat

DataComponentKind

DecoratedPotPattern

DecoratedPotPatterns

DimensionType

Enchantment

EnchantmentEffectComponentKind

EnchantmentEntityEffectKind

EnchantmentLevelBasedValueKind

EnchantmentLocationBasedEffectKind

EnchantmentProviderKind

EnchantmentValueEffectKind

EntityKind

An enum that contains every type of entity.

EntitySubPredicateKind

FloatProviderKind

Fluid

FrogVariant

GameEvent

HeightProviderKind

HolderSet

Instrument

IntProviderKind

Item

ItemSubPredicateKind

JukeboxSong

LootConditionKind

LootFunctionKind

LootNbtProviderKind

LootNumberProviderKind

LootPoolEntryKind

LootScoreProviderKind

MapDecorationKind

MemoryModuleKind

Menu

MenuKind

MobEffect

NumberFormatKind

PaintingVariant

ParticleKind

PointOfInterestKind

PosRuleTest

PositionSourceKind

Potion

RecipeBookCategory

RecipeDisplay

RecipeKind

RecipeSerializer

RuleBlockEntityModifier

RuleTest

Schedule

SensorKind

SlotDisplay

SoundEvent

StatKind

TriggerKind

TrimMaterial

TrimPattern

VillagerKind

VillagerProfession

WolfVariant

WorldgenBiomeSource

WorldgenBlockStateProviderKind

WorldgenCarver

WorldgenChunkGenerator

WorldgenDensityFunctionKind

WorldgenFeature

WorldgenFeatureSizeKind

WorldgenFoliagePlacerKind

WorldgenMaterialCondition

WorldgenMaterialRule

WorldgenPlacementModifierKind

WorldgenPoolAliasBinding

WorldgenRootPlacerKind

WorldgenStructureKind

WorldgenStructurePiece

WorldgenStructurePlacement

WorldgenStructurePoolElement

WorldgenStructureProcessor

WorldgenTreeDecoratorKind

WorldgenTrunkPlacerKind

Traits

Registry